Refactor: Move `renderField` in formUtilities
